pyproject.toml
[tool.poetry]
name = "bicing-utils"
version = "0.1.0"
description = ""
authors = ["gonzalogacc <gonzalogacc@gmail.com>"]
readme = "README.md"
packages = [{include = "bicing_utils"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
geopy = "^2.3.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
